requirements.txt
streamlit>=1.25.0
pandas>=1.5.0
numpy>=1.20.0
plotly>=5.10.0
matplotlib>=3.5.0
pillow>=9.0.0
openpyxl>=3.0.0 